Woah nelly.  Starting the merge, doing the sessions.
