+++ addon/data/browserid/channel.js
