+++ addon/data/browserid/injector.js
