+++ addon/data/sessions/page_interaction.js
