+++ addon/lib/browserid/helpers.js
