+++ addon/lib/browserid/main.js
