+++ addon/lib/browserid/window_listener.js
