+++ addon/lib/helpers.js
