+++ addon/lib/main.js
@@ -1,0 +1,7 @@
+const {MainSession} = require("sessions/main_module");
+
+exports.main = function() {
+  MainSession();
+
+}
+
