+++ addon/lib/sessions/bindings.js
