+++ addon/lib/sessions/cookie_monster.js
