+++ addon/lib/sessions/model.js
