+++ addon/lib/sessions/permissions.js
