+++ addon/lib/sessions/session_display.js
