+++ addon/lib/sessions/window_session.js
