+++ addon/test/browserid/test-main.js
